hal_can

Trait ViCan

pub trait ViCan {
    type Error: Debug;

    // Required methods
    fn configure(&mut self, kbps: u32) -> Result<(), Self::Error>;
    fn send_frame(&mut self, frame: &CanFrame) -> Result<(), Self::Error>;
    fn recv_frame(&mut self) -> Result<CanFrame, Self::Error>;
}

CAN bus controller trait.

Contract

All operations are synchronous.

Bit rate must be configured before any frame is sent or received.

After BusOff, the implementation must be reset before further use.

Required Associated Types

type Error: Debug

Required Methods

fn configure(&mut self, kbps: u32) -> Result<(), Self::Error>

Configure bit timing. Supported rates: 125, 250, 500, 1000 kbps. Returns InvalidBitrate for any other value.

fn send_frame(&mut self, frame: &CanFrame) -> Result<(), Self::Error>

Transmit a CAN frame. Returns TxFull if the transmit buffer is full.

fn recv_frame(&mut self) -> Result<CanFrame, Self::Error>

Receive the next available frame. Returns RxEmpty if no frame is pending.